requirements.txt
loguru
pydantic[dotenv]
aiogram
asyncpg
hurry.filesize